Set BuildActiveArchitecturesOnly to No for debug builds
This is necessary to allow us to build projects that have other external restrictions that do not allow us to build in the latest possible architechture. For example if we have a library that will only build in 32 bit, but we try to debug against a 64 bit device, the main project will build as a 32 bit binary, but the library will be 64 bit only and they will fail to link.